+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+1.9.11 / 2014-06-05
+===================
+
+* Fix first level transitive calculation, makes sure project dependencies do not have locked versions
+
+1.9.10 / 2014-06-04
+===================
+
+* Fix to commitLock tag
+* Add -PdependencyLock.useGeneratedLock flag
+
 1.9.9 / 2014-06-03
 ==================